+++ packages/gymui/src/components/Card/card-title.tsx
@@ -1,0 +1,14 @@
+import styles from './Card.module.css'
+
+export interface TitleProps extends React.HTMLAttributes<HTMLHeadingElement> {
+  children: React.ReactNode
+  sx?: React.CSSProperties
+}
+
+export const Title = ({ children, sx }: TitleProps) => {
+  return (
+    <p style={sx} className={styles.title}>
+      {children}
+    </p>
+  )
+}